requirements.txt
Django==2.1.7
django-ckeditor==5.6.1
django-jet==1.0.8
django-js-asset==1.2.2
django-pagedown==1.0.6
google-api-python-client==1.4.1
gunicorn==19.9.0
httplib2==0.12.1
oauth2client==4.1.3
Pillow==5.4.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
rsa==4.0
six==1.12.0
uritemplate==3.0.0
virtualenv==16.4.3
virtualenvwrapper-win==1.2.5
whitenoise==4.1.2
